+++ public/assignment/views/widget/controllers/widget-edit.controller.client.js
@@ -4,7 +4,7 @@
 (function () {
     angular
         .module("WebAppMaker")
-        .controller("EditWidgetController", editWidgetController);
+        .controller("EditWidgetController", editWidgetController)
 
     function editWidgetController($routeParams, $location, WidgetService) {
         var vm = this;
@@ -12,13 +12,18 @@
         vm.websiteID = $routeParams['wid'];
         vm.pageID = $routeParams['pid'];
         vm.widgetID = $routeParams['wgid'];
+        vm.url = "/api/upload?uid=" + vm.userID + "&wid=" + vm.websiteID +
+            "&pid=" + vm.pageID;
 
         vm.updateWidget = updateWidget;
         vm.deleteWidget = deleteWidget;
         vm.getEditorTemplateUrl = getEditorTemplateUrl;
 
         function init() {
-            vm.widget = WidgetService.findWidgetById(vm.widgetID);
+            WidgetService.findWidgetById(vm.widgetID)
+                .success(function (widget) {
+                    vm.widget = widget;
+                });
         }
 
         init();
@@ -27,19 +32,30 @@
             return 'views/widget/templates/editors/widget-' + type + '-editor.view.client.html';
         }
 
-        function deleteWidget () {
-            WidgetService.deleteWidget(vm.widgetID);
-            $location.url("/user/"+vm.userID+"/website/"+vm.websiteID+"/page/"+vm.pageID+"/widget");
+        function deleteWidget() {
+            var answer = confirm("Are you sure?");
+            if (answer) {
+                WidgetService.deleteWidget(vm.widgetID)
+                    .success(function () {
+                        $location.url("/user/" + vm.userID + "/website/" +
+                            vm.websiteID + "/page/" + vm.pageID + "/widget");
+                    })
+                    .error(function () {
+                        vm.error = 'Unable to delete widget';
+                    });
+            }
         }
 
-        function updateWidget (updatedWidget) {
-            var widget = WidgetService.updateWidget(vm.widgetID, updatedWidget);
-            if(widget == null) {
-                vm.error = "Unable to update widget";
-            } else {
-                vm.message = "Widget successfully updated";
-            }
-            console.log(widget);
+        function updateWidget(updatedWidget) {
+            WidgetService.updateWidget(vm.widgetID, updatedWidget)
+                .success(function () {
+                    vm.message = "Widget successfully updated";
+                    $location.url("/user/" + vm.userID + "/website/" +
+                        vm.websiteID + "/page/" + vm.pageID + "/widget");
+                })
+                .error(function () {
+                    vm.error = "Unable to update widget";
+                });
         }
     }
 })();
